-Added support for diagnostics -Source mapping overrides for contracts + functions in order to keep liner clean. -General cleanup
